Even closer, quick sort logic isn't fully correct but i have made a breakthrough
Seems to be there is also something wrong with the animation render as well. The swap logic has shown to have inconsistencies in terms of ordering some of the elements within the array, i.e. there is a significantly smaller value at the very end of the sorted array where it should be at the beginning with the other smaller values.